refactor(policies): move the token selector under Policies/SpendingLimitFlow
The selector, its data hook, merge/search utils and the popular-address table have no
consumer inside features/spending-limits; their only consumer is the WA-3150 form in
features/spaces/components/Policies, and ESLint forbids cross-feature components/** and
hooks/** imports. Living under Policies also keeps the Space flow gated by POLICIES alone.
